Create credentials module
