+++ src/strategies/podman_strategy.rs
@@ -1,9 +1,9 @@
 use std::process::{Command, Stdio};
 
-use anyhow::{anyhow, bail, Result};
+use anyhow::{bail, Result};
 use log::{debug, info, trace};
 
-use crate::{image_inspection::ImageInspection, strategies::ENV_CREDENTIALS};
+use crate::{credentials, image_inspection::ImageInspection};
 
 use super::{BuildStrategy, InspectStrategy};
 
@@ -59,16 +59,13 @@ impl BuildStrategy for PodmanStrategy {
     }
 
     fn login(&self) -> Result<()> {
-        let (registry, username, password) = ENV_CREDENTIALS
-            .as_ref()
-            .map(|credentials| {
-                (
-                    &credentials.registry,
-                    &credentials.username,
-                    &credentials.password,
-                )
-            })
-            .ok_or_else(|| anyhow!("Unable to login, missing credentials!"))?;
+        let (registry, username, password) = credentials::get_credentials().map(|credentials| {
+            (
+                &credentials.registry,
+                &credentials.username,
+                &credentials.password,
+            )
+        })?;
 
         trace!("podman login -u {username} -p [MASKED] {registry}");
         let output = Command::new("podman")
